% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_player_game_windows.R
\name{add_player_game_windows}
\alias{add_player_game_windows}
\title{Add player-game window functions to gamelogs}
\usage{
add_player_game_windows(player_gamelogs, team_gamelogs)
}
\arguments{
\item{player_gamelogs}{data.frame from \code{make_player_game_summaries()}}
}
\value{
data.frame
}
\description{
Add player-game window functions to gamelogs
}
