% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_data.R
\name{make_model_data}
\alias{make_model_data}
\title{Combine inputs and outcomes and split for modeling}
\usage{
make_model_data(team_model_inputs, team_model_outcomes, OUTCOME)
}
\arguments{
\item{team_model_inputs}{data.frame}

\item{team_model_outcomes}{data.frame}

\item{OUTCOME}{UNQUOTED name of outcome for this run}
}
\value{
list
}
\description{
Combine inputs and outcomes and split for modeling
}
