% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_player_game_summaries.R
\name{make_player_game_summaries}
\alias{make_player_game_summaries}
\title{Make player game summaries that are a bit more enriched than the gamelogs}
\usage{
make_player_game_summaries(
  gamelogs,
  team_gamelogs,
  cols_to_rename = c(get_predictor_vec(), get_exclusion_vec())
)
}
\arguments{
\item{gamelogs}{data.frame output from \code{prep_gamelogs()}}

\item{team_gamelogs}{data.frame output from \code{make_team_game_summaries()}}

\item{cols_to_rename}{character of columns that need to be labeled as 'player' fields}
}
\value{
data.frame
}
\description{
Make player game summaries that are a bit more enriched than the gamelogs
}
\note{
this is the first summary step for players, look for \code{app_player_game_windows()} and other steps later in the pipeline
}
