% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_team_game_metadata.R
\name{make_team_game_metadata}
\alias{make_team_game_metadata}
\title{Make metadata for team-games}
\usage{
make_team_game_metadata(gamelogs)
}
\arguments{
\item{gamelogs}{data.frame from \code{prep_gamelogs()}}
}
\value{
data.frame
}
\description{
Make metadata for team-games
}
\note{
this should be joinable on modeling data without any additional lagging calculations, i.e. all data should be in scope for predictions unless otherwise stated (e.g. outcomes)
}
