% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_team_game_recipe.R
\name{make_team_game_recipe}
\alias{make_team_game_recipe}
\title{Make a tidymodels recipe for building team-game models}
\usage{
make_team_game_recipe(team_modeling_data)
}
\arguments{
\item{team_modeling_data}{modeling data from \code{make_model_data()}}
}
\value{
a recipe
}
\description{
Make a tidymodels recipe for building team-game models
}
