% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_team_game_summaries.R
\name{make_team_game_summaries}
\alias{make_team_game_summaries}
\title{Summarise gamelogs to make team game data}
\usage{
make_team_game_summaries(
  gamelogs,
  include_vec = get_predictor_vec(),
  exclude_vec = get_exclusion_vec()
)
}
\arguments{
\item{gamelogs}{data.frame from \code{prep_gamelogs()}}

\item{include_vec}{character vector of field to include for aggregation, from \code{get_predictor_vec()}}

\item{exclude_vec}{character vector of field to include for aggregation, from \code{get_exclusion_vec()}}
}
\value{
data.frame
}
\description{
Summarise gamelogs to make team game data
}
\note{
This won't include any window functions, just team-game totals of stuff in the \code{include_vec} argument
}
