% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.pipeline-package.R
\docType{package}
\name{nba.pipeline-package}
\alias{nba.pipeline}
\alias{nba.pipeline-package}
\title{nba.pipeline: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jim Kloet \email{jim.kloet@gmail.com}

}
\keyword{internal}
