% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_gamelogs.R
\name{prep_gamelogs}
\alias{prep_gamelogs}
\title{prep gamelogs for aggregatin'}
\usage{
prep_gamelogs(path)
}
\arguments{
\item{path}{where are the gamelogs? they oughta be csv too else this won't work}
}
\value{
data.frame
}
\description{
prep gamelogs for aggregatin'
}
