% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_team_model_inputs.R
\name{prep_team_model_inputs}
\alias{prep_team_model_inputs}
\title{Construct initial input for team game models}
\usage{
prep_team_model_inputs(
  team_gamelogs,
  team_game_metadata,
  lag_regex = "_team_(rolling|allowed|diff)|window_rolling",
  rolling_regex = "rolling_(sum|sd|mean)$"
)
}
\arguments{
\item{team_gamelogs}{data.frame}

\item{team_game_metadata}{data.frame}

\item{lag_regex}{character}

\item{rolling_regex}{character}
}
\value{
data.frame
}
\description{
Construct initial input for team game models
}
