% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cumvar}
\alias{cumvar}
\title{cumulative variance}
\usage{
cumvar(x, sd = FALSE)
}
\arguments{
\item{x}{numeric}

\item{sd}{logical should it return standard deviation instead of variance}
}
\value{
numeric
}
\description{
cumulative variance
}
