% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_team_game_windows.R
\name{add_team_game_windows}
\alias{add_team_game_windows}
\title{Add window functions to team-game summary data}
\usage{
add_team_game_windows(team_gamelogs)
}
\arguments{
\item{team_gamelogs}{data.frame of summarised team-game data}
}
\value{
data.frame
}
\description{
Add window functions to team-game summary data
}
\note{
Most of the windows are going to be within season-types, i.e. pre-season, regular, and post-season. For now.
}
